# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Manajemen Tiket Tenant API",
        "description": "Dokumentasi API Terintegrasi untuk Manajemen Tiket Keluhan Properti (Dawai)",
        "contact": {
            "name": "Contact the developer"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Main Server Host Lokal"
        }
    ],
    "paths": {
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Mengambil daftar tiket (IAE-T2 Standard)",
                "operationId": "getTicketsList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Menambahkan tiket keluhan baru",
                "operationId": "storeTicket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "listing_id",
                                    "contract_id",
                                    "tenant_name",
                                    "tenant_email",
                                    "description"
                                ],
                                "properties": {
                                    "listing_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "contract_id": {
                                        "type": "string",
                                        "example": "100"
                                    },
                                    "tenant_name": {
                                        "type": "string",
                                        "example": "Dawai"
                                    },
                                    "tenant_email": {
                                        "type": "string",
                                        "example": "dawai@telkom.id"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "AC Bocor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tiket berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Mengambil detail spesifik tiket",
                "operationId": "getTicketById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menemukan tiket",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan NIM Anda (102022400198) untuk membuka akses API",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Tickets",
            "description": "Tickets"
        }
    ]
}